Rename distribution components
