+++ Tests/RoktUXHelperTests/UI/Components/TestCarouselDistributionComponent.swift
@@ -1,5 +1,5 @@
 //
-//  TestOneByOneComponent.swift
+//  TestCarouselDistributionComponent.swift
 //  RoktUXHelperTests
 //
 //  Licensed under the Rokt Software Development Kit (SDK) Terms of Use
@@ -14,12 +14,11 @@ import SwiftUI
 import ViewInspector
 @testable import RoktUXHelper
 import DcuiSchema
-import SnapshotTesting
 
 @available(iOS 15.0, *)
-final class TestOneByOneComponent: XCTestCase {
+final class TestCarouselDistributionComponent: XCTestCase {
 
-    func test_one_by_one() throws {
+    func test_carousel() throws {
         var closeActionCalled = false
         let view = try TestPlaceHolder.make(
             eventHandler: { event in
@@ -27,82 +26,63 @@ final class TestOneByOneComponent: XCTestCase {
                     closeActionCalled = true
                 }
             },
-            layoutMaker: LayoutSchemaViewModel.makeOneByOne(layoutState:eventService:)
+            layoutMaker: LayoutSchemaViewModel.makeCarousel(layoutState:eventService:)
         )
-        let oneByOneComponent = try view.inspect().view(TestPlaceHolder.self)
+
+        let carouselComponent = try view.inspect().view(TestPlaceHolder.self)
             .view(EmbeddedComponent.self)
             .vStack()[0]
             .view(LayoutSchemaComponent.self)
-            .view(OneByOneComponent.self)
+            .view(CarouselDistributionComponent.self)
             .actualView()
-
-        let group = try oneByOneComponent
+        
+        let carousel = try carouselComponent
             .inspect()
-            .group()
-
-        let oneByOne = try group
             .find(LayoutSchemaComponent.self)
-
+        
         // test custom modifier class
-        let paddingModifier = try oneByOne.modifier(PaddingModifier.self)
+        let paddingModifier = try carousel.modifier(PaddingModifier.self)
         XCTAssertEqual(try paddingModifier.actualView().padding, FrameAlignmentProperty(top: 3, right: 4, bottom: 5, left: 6))
-
+        
         // test the effect of custom modifier
-        let padding = try oneByOne.padding()
+        let padding = try carousel.padding()
         XCTAssertEqual(padding, EdgeInsets(top: 3.0, leading: 6.0, bottom: 5.0, trailing: 4.0))
+        
+        XCTAssertEqual(try carousel.accessibilityLabel().string(), "Page 1 of 1")
 
-        XCTAssertEqual(try group.accessibilityLabel().string(), "Offer 1 of 1")
-
-        oneByOneComponent.goToNextOffer()
+        carouselComponent.goToNextOffer()
         XCTAssertTrue(closeActionCalled)
     }
-
+    
     func test_goToNextOffer_with_closeOnComplete_false() throws {
         var closeActionCalled = false
-        var SignalResponseCalled = false
-
         let closeOnCompleteSettings = LayoutSettings(closeOnComplete: false)
-
         let view = try TestPlaceHolder.make(
             layoutSettings: closeOnCompleteSettings,
             eventHandler: { event in
                 if event.eventType == .SignalDismissal {
                     closeActionCalled = true
-                } else if event.eventType == .SignalResponse {
-                    SignalResponseCalled = true
                 }
             },
-            layoutMaker: LayoutSchemaViewModel.makeOneByOne(layoutState:eventService:)
+            layoutMaker: LayoutSchemaViewModel.makeCarousel(layoutState:eventService:)
         )
-        let oneByOneComponent = try view.inspect().view(TestPlaceHolder.self)
+        
+        let carouselComponent = try view.inspect().view(TestPlaceHolder.self)
             .view(EmbeddedComponent.self)
             .vStack()[0]
             .view(LayoutSchemaComponent.self)
-            .view(OneByOneComponent.self)
+            .view(CarouselDistributionComponent.self)
             .actualView()
 
-        XCTAssertFalse(SignalResponseCalled)
-
-        oneByOneComponent.goToNextOffer()
+        carouselComponent.goToNextOffer()
         XCTAssertFalse(closeActionCalled)
-        XCTAssertFalse(SignalResponseCalled)
-    }
-
-    func testEmbeddedOneByOne() {
-        // Create a RoktLayoutUIView with TestViewController
-        // TODO: investigate perceptual precision usage.
-//        withSnapshotTesting(diffTool: .ksdiff) {
-//            waitForViewController("embedded_onebyone") { testViewController in
-//                assertSnapshot(of: testViewController, as: .image(on: .iPhone13Pro(.portrait)))
-//            }
-//        }
     }
 }
 
 @available(iOS 15.0, *)
 extension LayoutSchemaViewModel {
 
-    static func makeOneByOne(
+    static func makeCarousel(
         layoutState: LayoutState,
         eventService: EventService
     ) throws -> Self {
@@ -110,10 +90,7 @@ extension LayoutSchemaViewModel {
         let transformer = LayoutTransformer(layoutPlugin: get_mock_layout_plugin(slots: slots!),
                                             layoutState: layoutState,
                                             eventService: eventService)
-        let model = ModelTestData.OneByOneData.oneByOne()
-        return LayoutSchemaViewModel.oneByOne(try transformer.getOneByOne(
-            oneByOneModel: model!,
-            context: .outer(slots!.map(\.offer))
-        ))
+        let model = ModelTestData.CarouselData.carousel()
+        return LayoutSchemaViewModel.carousel(try transformer.getCarousel(carouselModel: model!, context: .outer([])))
     }
 }
